refactor(transport-browser-windows): split runtime modules for code-principles compliance
